// frontend/src/app/shared/enums/vehicle/vehicle-category.enum.ts
export enum VehicleCategory {
  TOUT_TERRAIN = 'Tout terrain',
  BERLINE_TAILLE_L = 'Berline taille L',
  BERLINE_TAILLE_M = 'Berline taille M',
  SUV = 'SUV',
  MINI_CITADINE = 'Mini citadine',
  MICRO_URBAINE = 'Micro urbaine',
  PICKUP = 'Pickup',
  CITADINE_POLYVALENTE = 'Citadine polyvalente',
  COMPACTE = 'Compacte',
  BERLINE_TAILLE_S = 'Berline taille S'
}